codelist {countrycode}	R Documentation

Country Code Translation Data Frame (Cross-Sectional)

Description

A data frame used internally by the countrycode() function. 'countrycode' can use any valid code as destination, but only some codes can be used as origin.

Format

data frame with codes as columns

Details

Origin and Destination:

country.name: country name (English)

country.name.de: country name (German)

cowc: Correlates of War character

cown: Correlates of War numeric

ecb: European Central Bank

eurostat: Eurostat

fao: Food and Agriculture Organization of the United Nations numerical code

fips: FIPS 10-4 (Federal Information Processing Standard)

gaul: Global Administrative Unit Layers

genc2c: GENC 2-letter code

genc3c: GENC 3-letter code

genc3n: GENC numeric code

imf: International Monetary Fund

ioc: International Olympic Committee

iso2c: ISO-2 character

iso3c: ISO-3 character

iso2n: ISO-2 numeric

iso3n: ISO-3 numeric

p4n: Polity IV numeric country code

p4c: Polity IV character country code

un: United Nations numerical code

unpd: United Nations Procurement Division

wb: World Bank (very similar but not identical to iso3c)

wvs: World Values Survey numeric code

Destination only:

ar5: IPCC's regional mapping used both in the Fifth Assessment Report (AR5) and for the Reference Concentration Pathways (RCP)

continent: Continent as defined in the World Bank Development Indicators

cow.name: Correlates of War country name

ecb.name: European Central Bank country name

eurocontrol_pru: European Organisation for the Safety of Air Navigation

eurocontrol_statfor: European Organisation for the Safety of Air Navigation

eurostat.name: Eurostat country name

eu28: Member states of the European Union (as of December 2015), without special territories

fao.name: Food and Agriculture Organization of the United Nations country name

fips.name: FIPS 10-4 Country name

genc.name: Geopolitical Entities, Names and Codes standard country names

icao: International Civil Aviation Organization

icao_region: International Civil Aviation Organization (Region)

ioc.name: International Olympic Committee country name

iso.name.en: ISO English short name

iso.name.fr: ISO French short name

p4.name: Polity IV country name

region: Regions as defined in the World Bank Development Indicators

un.name.ar: United Nations Arabic country name

un.name.en: United Nations English country name

un.name.es: United Nations Spanish country name

un.name.fr: United Nations French country name

un.name.ru: United Nations Russian country name

un.name.zh: United Nations Chinese country name

unpd.name: United Nations Procurement Division country name

wvs.name: World Values Survey numeric code country name

cldr.*: 622 country name variants from the UNICODE CLDR project. Inspect the 'countrycode::cldr_examples' data.frame for a full list of available country names and examples.

Note

The Correlates of War (cow) and Polity 4 (p4) project produce codes in country year format. Some countries go through political transitions that justify changing codes over time. When building a purely cross-sectional conversion dictionary, this forces us to make arbitrary choices with respect to some entities (e.g., Western Germany, Vietnam, Serbia). 'countrycode' includes a reconciled dataset in panel format: 'countrycode::countrycode_panel'. Instead of converting code, we recommend that users dealing with panel data "left-merge" their data into this panel dictionary.

[Package countrycode version 1.00.0 Index]