countrycode-package {countrycode}	R Documentation

Convert Country Codes or Country Names

Description

Convert country codes or country names

Details

The countrycode function can convert to and from several different country coding schemes. It uses regular expressions to convert country names (e.g. Sri Lanka) into any of those coding schemes, or into standardized country names in several languages. It can create variables with the name of the continent and/or several regional groupings to which each country belongs.

Type ?codelist to get a list of available origin and destination codes.

Author(s)

Vincent Arel-Bundock vincent.arel-bundock@umontreal.ca

References

http://arelbundock.com https://github.com/vincentarelbundock/countrycode

[Package countrycode version 1.00.0 Index]