countryname_dict {countrycode}	R Documentation

A dataframe of alternative country names in many languages. Used internally by the countryname function.

Description

A dataframe of alternative country names in many languages. Used internally by the countryname function.

Format

dataframe

[Package countrycode version 1.7.0 Index]