get_dictionary {countrycode}	R Documentation

Get Custom Dictionaries

Description

Download a custom dictionary to use in the custom_dict argument of countrycode()

Usage

get_dictionary(dictionary = NULL)

Arguments

dictionary
A character string that specifies the dictionary to be retrieved. It must be one of "global_burden_of_disease", "ch_cantons", "us_states", "exiobase3", "gtap10". If NULL, the function will print the list of available dictionaries. Default is NULL.

Value

If a valid dictionary is specified, the function will return that dictionary as a data.frame. If an invalid dictionary or no dictionary is specified, the function will stop and throw an error message.

Examples

## Not run: 
cd <- get_dictionary("us_states")
countrycode::countrycode(c("MO", "MN"), origin = "state.abb", "state.name", custom_dict = cd)

## End(Not run)

[Package countrycode version 1.7.0 Index]